move extended bounds rounding to date histo agg builder
